+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to IG Publisher Manager will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.5.0] - 2025-08-04
+## [v0.6.0] - 2025-08-04
+### Added
+- Tools... Publish to website and related changes. 
+
+### Changed
+- Add icons to all menus
+
+## [v0.5.0] - 2025-08-04
 ### Added
 - Documentation button with dropdown menu for quick access to FHIR IG resources